==> 1-4. Preprocessing.ipynb <==
# 데이터 전처리(Literacy)

---

1. 개봉일 데이터와 merge
2. 흥행영화 추출
3. 배급사(국내, 상장, 주요) 더미화, 연령 더미화
4. 배우, 감독 가중치 파생변수 생성

## 라이브러리 및 데이터 로드

## 라이브러리 로드
import pandas as pd
import statsmodels.api as sm
import numpy as np
import matplotlib.pyplot as plt
import os

# 경고 메세지 삭제
import warnings
warnings.filterwarnings('ignore')

# 한글 깨짐 방지
import matplotlib.pyplot as plt
plt.rcParams['font.family'] = 'Malgun Gothic'

# pd.set_option('display.max_rows', None)

---
# 1. 개봉일 데이터와 merge

## 데이터 로드
df = pd.read_excel('./datasets/movie_1.xlsx')
df

df['영화명'] = df['영화명'].astype('object')

# index 0번 삭제 및 순번 컬럼 삭제
df.drop('순번', axis=1, inplace=True, errors='ignore')
df.drop(index=0, inplace=True)

# 컬럼명 변경
df.rename(columns={'전국':'전국 스크린수', '전국.1':'전국 매출액', '전국.2': '전국 관객수', '서울':'서울 매출액', '서울.1':'서울 관객수' },
           inplace=True, errors='ignore')

## 영화명이 동일한 데이터 구분

# 나를 찾아줘 영화 구분
df[(df['영화명'] == '나를 찾아줘') & (df['개봉일'] == '2014-10-23')]
df.loc[(df['영화명'] == '나를 찾아줘') & (df['개봉일'] == '2014-10-23'), '영화명'] = '나를 찾아줘(외)'

# 스파이 영화 구분
df[(df['영화명'] == '스파이') & (df['개봉일'] == '2015-05-21')] 
df.loc[(df['영화명'] == '스파이') & (df['개봉일'] == '2015-05-21'), '영화명'] = '스파이(외)'

# 미녀와 야수1 영화 구분
df[(df['영화명'] == '미녀와 야수') & (df['개봉일'] == '2014-06-18')] 
df.loc[(df['영화명'] == '미녀와 야수') & (df['개봉일'] == '2014-06-18'), '영화명'] = '미녀와 야수1'

# 미녀와 야수2 영화 구분
df[(df['영화명'] == '미녀와 야수') & (df['개봉일'] == '2017-03-16')]
df.loc[(df['영화명'] == '미녀와 야수') & (df['개봉일'] == '2017-03-16'), '영화명'] = '미녀와 야수2'

## 개봉일 데이터 로드
daily_df = pd.read_csv('./datasets/Movie_Daily_Data.csv',index_col=0)
daily_df

# 개봉일 2010년 이후 데이터만 추출
daily_df = daily_df[daily_df['개봉일'] > '2010-01-01']

daily_df['영화명'] = daily_df['영화명'].astype('object')

daily_df = daily_df[daily_df['해당일'] == daily_df['개봉일']]

daily_df['개봉일'] = pd.to_datetime(daily_df['개봉일'])

## 영화명이 동일한 데이터 구분

# 나를 찾아줘 영화 구분
daily_df[(daily_df['영화명'] == '나를 찾아줘') & (daily_df['개봉일'] == '2014-10-23')]
daily_df.loc[(daily_df['영화명'] == '나를 찾아줘') & (daily_df['개봉일'] == '2014-10-23'), '영화명'] = '나를 찾아줘(외)'

# 스파이 영화 구분
daily_df[(daily_df['영화명'] == '스파이') & (daily_df['개봉일'] == '2015-05-21')] 
daily_df.loc[(daily_df['영화명'] == '스파이') & (daily_df['개봉일'] == '2015-05-21'), '영화명'] = '스파이(외)'

# 미녀와 야수1 영화 구분
daily_df[(daily_df['영화명'] == '미녀와 야수') & (daily_df['개봉일'] == '2014-06-18')] 
daily_df.loc[(daily_df['영화명'] == '미녀와 야수') & (daily_df['개봉일'] == '2014-06-18'), '영화명'] = '미녀와 야수1'

# 미녀와 야수2 영화 구분
daily_df[(daily_df['영화명'] == '미녀와 야수') & (daily_df['개봉일'] == '2017-03-16')]
daily_df.loc[(daily_df['영화명'] == '미녀와 야수') & (daily_df['개봉일'] == '2017-03-16'), '영화명'] = '미녀와 야수2'

daily_df['영화명'] = daily_df['영화명'].str.strip()

# 데이터 결합
df = pd.merge(df, daily_df, how='left', on=['영화명', '개봉일'])

---
# 2. 흥행영화 추출

# 연도만 추출하여 새로운 파생변수 개봉연도 생성
df['개봉일'] = pd.to_datetime(df['개봉일'])
df['개봉연도'] = df['개봉일'].dt.year

# 새로운 파생변수 평균 제작비 컬럼 생성 ( 한국영화진흥원 보고서 참고 )
def change(x):
    if x == 2009:
        result = 23.1
    elif x == 2010:
        result = 21.6
    elif x == 2011:
        result = 22.7
    elif x == 2012:
        result = 20.3
    elif x == 2013:
        result = 21.4
    elif x == 2014:
        result = 20.1
    elif x == 2015:
        result = 19.9
    elif x == 2016:
        result = 24.0
    elif x == 2017:
        result = 26.3
    elif x == 2018:
        result = 26.8
    elif x == 2019:
        result = 29.0
    else:
        result = '!'
    return result

df['평균 제작비'] = df['개봉연도'].apply(change)

# 매출액 억단위로 변경
df['수정전국매출액']  =  df['전국 매출액'] / 100000000

# 평균 제작비보다 매출액이 더 높은 데이터만 추출
df = df[df['수정전국매출액'] >= df['평균 제작비']]

# 영화명이 2개인 데이터 확인 -> 데이터 merge과정에서 중복 결합 확인 -> 모두 제거
df['영화명'].value_counts().head(14)

# 중복 영화 데이터 제거
df = df.drop_duplicates('영화명')

# 인덱스 초기화
df.reset_index(drop=True, inplace=True)

df.columns

df = df[['영화명', '배급사', '개봉일', '국적', '전국 스크린수', '전국 매출액', '전국 관객수', '서울 매출액', '서울 관객수', 
    '장르', '등급', '영화구분', '개봉연도', '해당일의 관객수', '당일 매출액', '해당일자 매출총액 대비 해당 영화의 매출비율', '대표코드']]

df[df['대표코드'].isna()]

# 결측치 채우기
df.loc[df['영화명'] == 1987, '해당일의 관객수'] = 331736
df.loc[df['영화명'] == 1987, '당일 매출액'] = 1973650400
df.loc[df['영화명'] == 1987, '해당일자 매출총액 대비 해당 영화의 매출비율'] = 32.3
df.loc[df['영화명'] == 1987, '대표코드'] = 20170590

df.loc[df['영화명'] == '울지마, 톤즈', '해당일의 관객수'] = 2666
df.loc[df['영화명'] == '울지마, 톤즈', '당일 매출액'] = 17689500
df.loc[df['영화명'] == '울지마, 톤즈', '해당일자 매출총액 대비 해당 영화의 매출비율'] = 0.9
df.loc[df['영화명'] == '울지마, 톤즈', '대표코드'] = 20100962

df.loc[df['영화명'] == '내 이름은 칸', '해당일의 관객수'] = 14975
df.loc[df['영화명'] == '내 이름은 칸', '당일 매출액'] = 119445500
df.loc[df['영화명'] == '내 이름은 칸', '해당일자 매출총액 대비 해당 영화의 매출비율'] = 3.2
df.loc[df['영화명'] == '내 이름은 칸', '대표코드'] = 20119771

df.loc[df['영화명'] == '극장판 짱구는 못말려: 아뵤! 쿵후 보이즈 ~라면 대란~', '해당일의 관객수'] = 44580
df.loc[df['영화명'] == '극장판 짱구는 못말려: 아뵤! 쿵후 보이즈 ~라면 대란~', '당일 매출액'] = 348391550
df.loc[df['영화명'] == '극장판 짱구는 못말려: 아뵤! 쿵후 보이즈 ~라면 대란~', '해당일자 매출총액 대비 해당 영화의 매출비율'] = 3.3
df.loc[df['영화명'] == '극장판 짱구는 못말려: 아뵤! 쿵후 보이즈 ~라면 대란~', '대표코드'] = 20183479

# 대표코드가 없는 데이터 제거
df = df[~df['대표코드'].isna()]

df.reset_index(drop=True, inplace=True)

df['대표코드'] = df['대표코드'].astype(int)

# df.to_csv('./datasets/배우감독파생변수용.csv')

---
# 3. 더미화

### 국적, 영화등급, 영화구분 더미화

## 국적 컬럼 변경 (국내 : 1, 해외 : 0)

nation = {
    "한국": 1,
}

df['국적'] = df['국적'].map(nation).fillna(0)

df['국적'] = df['국적'].astype(int)

## 영화등급 컬럼 변경

grade = {
    "청소년관람불가": '청불',
    "15세이상관람가":  '15',
    "12세이상관람가":  '12',
    "전체관람가": '전체',
    "청소년관람불가,15세이상관람가" : '청불'
}

df['등급'] = df['등급'].map(grade)

dummies_genre = pd.get_dummies(df['등급'], prefix='등급')
df = pd.concat([df,dummies_genre], axis=1)

## 영화구분 분류 (1 : 일반영화, 0 : 독립/예술영화)

m = {
    "일반영화": 1,
    "독립/예술영화": 0
}

df['영화구분'] = df['영화구분'].map(m)

### 배급사 더미화
* 상장배급사 (상장 : 1, 비상장 : 0)
* 국내배급사 (국내 : 1, 해외 : 0)
* 주요배급사 (메이저 : 1, 비메이저 : 0)

# 배급사 컬럼만 추출
gg = df['배급사']

# 배급사 컬럼에서 외국 영화사명 포함된 values 추출 및 병합
a1 = gg[gg.str.contains("디즈니")]
a2 = gg[gg.str.contains("워너브러더스")]
a3 = gg[gg.str.contains("폭스")]
a4 = gg[gg.str.contains("소니")]
a5 = gg[gg.str.contains("유니버설")]
kkk = pd.concat([a1, a2, a3, a4, a5], axis=0)

# 새로 생성된 컬럼 범주화(0,1)

for i in kkk:
    for j in range(len(df)):
        if i in df['배급사'][j]:
            df.loc[j, '국내배급사'] = 1

ttt = df['국내배급사']
ttt.fillna(0, inplace=True)

# 배급사 컬럼에서 상장된 회사 추출 및 병합
b1 = gg[gg.str.contains("씨제이")]
b2 = gg[gg.str.contains("CJ")]
b3 = gg[gg.str.contains("CGV")]
b4 = gg[gg.str.contains("롯데")]
b5 = gg[gg.str.contains("NEW")]
b6 = gg[gg.str.contains("쇼박스")]
b7 = gg[gg.str.contains("스튜디오미르")]
b8 = gg[gg.str.contains("위지윅스튜디오")]
b9 = gg[gg.str.contains("덱스터")]
b10 = gg[gg.str.contains("samg엔터")]
b11= gg[gg.str.contains("애니플러스")]
b12 = gg[gg.str.contains("바른손이앤에이")]
b13 = gg[gg.str.contains("판타지오")]
b14= gg[gg.str.contains("스튜디오산타클로스")]
b15= gg[gg.str.contains("아센디오")]
b16 = gg[gg.str.contains("IHQ")]
b17= gg[gg.str.contains("콘텐트리중앙")]

kkkk = pd.concat([b1, b2, b3, b4, b5, b6, b7, b8, b9, b10, b11, b12, b13, b14, b15, b16, b17], axis=0)

# 새로 생성된 컬럼 범주화(0,1)

for i in kkkk:
    for j in range(len(df)):
        if i in df['배급사'][j]:
            df.loc[j, '상장배급사'] = 1

df['상장배급사'].fillna(0, inplace=True)
ttttt = df['상장배급사']

# 배급사(메이저 비메이저) 컬럼 0,1로 변환(바이너리)후 새 컬럼에 삽입
df['배급사'] = df['배급사'].fillna('aa')

ls = ['이엔엠', 'CJ', '롯데', '쇼박스', '판다', '메가박스', '워너', '유니버설', '소니픽', '디즈니', '판씨', '키다리', 'NEW', '엣나인', '이수', '리틀', '스마일이', '팝엔', '오퍼스', '코리아스', '이십세기', '올스타']

for i in ls:
    for j in range(len(df)):
        if i in df['배급사'][j]:
            df.loc[:, '배급사'][j] = 'O'
            
df['주요배급사'] = df['배급사'].map(lambda x: 1 if x == 'O' else 0)

df['주요배급사'].value_counts()

# 배급사 컬럼 삭제
df = df.drop('배급사', axis=1)

---
# 4. 파생변수생성

### 배우가중치 
* (배우출현횟수 / 관객수가중치)
### 감독가중치
* (감독출현횟수 / 관객수가중치)

actor = pd.read_csv('./datasets/배우가중치.csv', index_col=0)
directors = pd.read_csv('./datasets/감독가중치.csv', index_col=0)

df = pd.concat([df, actor, directors], axis=1)

df['국내배급사'] = df['국내배급사'].astype(int)
df['상장배급사'] = df['상장배급사'].astype(int)
df['주요배급사'] = df['주요배급사'].astype(int)

df['배우가중치'] = round(df['배우가중치'],2)
df['감독가중치'] = round(df['감독가중치'],2)

df.columns

df.columns = ['영화명', '개봉일', '국적', '전국스크린수', '전국매출액', '전국관객수', '서울매출액', '서울관객수',
       '장르', '등급', '영화구분', '개봉연도', '개봉일관객수', '개봉일매출액',
       '개봉일매출비율', '대표코드', '등급_12', '등급_15', '등급_전체', '등급_청불',
       '국내배급사', '상장배급사', '주요배급사', '배우가중치', '감독가중치']

df.drop(columns=['장르', '등급'], axis=1, inplace=True)

df

df.to_csv('./datasets/1. 전처리완료.csv')
