==> 3. Scaling.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import matplotlib.ticker as ticker
import seaborn as sns
from datetime import datetime
import time

from sklearn.model_selection import train_test_split
from sklearn.preprocessing import MinMaxScaler, StandardScaler, RobustScaler
from sklearn.metrics import accuracy_score,f1_score,recall_score,precision_score,confusion_matrix,ConfusionMatrixDisplay,roc_curve,roc_auc_score,precision_recall_curve
from sklearn.ensemble import RandomForestClassifier , StackingClassifier
from sklearn.svm import SVC
from sklearn.discriminant_analysis import LinearDiscriminantAnalysis
from sklearn.tree import DecisionTreeClassifier
from xgboost import XGBClassifier
from lightgbm import LGBMClassifier
from sklearn.neighbors import KNeighborsClassifier
from sklearn.linear_model import LogisticRegression,Lasso
from sklearn.preprocessing import Binarizer
from sklearn.model_selection import cross_val_score,GridSearchCV

# 한글 깨짐 방지
import matplotlib.pyplot as plt
plt.rcParams['font.family'] = 'malgun Gothic'

# 경고창 무시 
import warnings
warnings.filterwarnings("ignore")

df = pd.read_csv('./datasets/2. Labeling완료.csv', index_col=0)
df

---
# Train Test Set 분리

X = df.drop(columns='Label', axis=1)
y = df[['Label']]

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.3, random_state=0)

X_train_df = X_train[['전국스크린수', '전국매출액', '전국관객수', '서울매출액', '서울관객수', 
                   '개봉일관객수', '개봉일매출액', '개봉일매출비율', '배우가중치', '감독가중치']]
X_train_concat = X_train.drop(columns=['전국스크린수', '전국매출액', '전국관객수', '서울매출액', '서울관객수', 
                                       '개봉일관객수', '개봉일매출액', '개봉일매출비율', '배우가중치', '감독가중치'], axis=1)

X_test_df = X_test[['전국스크린수', '전국매출액', '전국관객수', '서울매출액', '서울관객수', 
                   '개봉일관객수', '개봉일매출액', '개봉일매출비율', '배우가중치', '감독가중치']]
X_test_concat = X_test.drop(columns=['전국스크린수', '전국매출액', '전국관객수', '서울매출액', '서울관객수', 
                                       '개봉일관객수', '개봉일매출액', '개봉일매출비율', '배우가중치', '감독가중치'], axis=1)

X_train_concat.reset_index(drop=True, inplace=True)
X_test_concat.reset_index(drop=True, inplace=True)
y_train.reset_index(drop=True, inplace=True)
y_test.reset_index(drop=True, inplace=True)

scaler_SS = StandardScaler()
train_scaled_SS = scaler_SS.fit_transform(X_train_df)
# transform()시 scale 변환된 데이터 셋이 numpy ndarry로 반환되어 이를 DataFrame으로 변환
X_train = pd.DataFrame(data = train_scaled_SS, columns=X_train_df.columns)

test_scaled_SS = scaler_SS.transform(X_test_df)
# transform()시 scale 변환된 데이터 셋이 numpy ndarry로 반환되어 이를 DataFrame으로 변환
X_test = pd.DataFrame(data = test_scaled_SS, columns=X_test_df.columns)

# Scaling Data와 결합

train_SS = pd.concat([X_train, X_train_concat, y_train], axis=1)
test_SS = pd.concat([X_test, X_test_concat, y_test], axis=1)

train_SS.to_csv('./datasets/train_ss.csv')
test_SS.to_csv('./datasets/test_ss.csv')
